fix: i delete infrastructureContentCard and put all the code in infrastructureCard
